+++ src/Hellion.Cluster/ClusterServer.cs
@@ -11,6 +11,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
+using Hellion.Core.Helpers;
 
 namespace Hellion.Cluster
 {
@@ -95,8 +96,7 @@ namespace Hellion.Cluster
         {
             Log.Info("New client connected from {0}", client.Socket.RemoteEndPoint.ToString());
 
-            if (client is ClusterClient)
-                (client as ClusterClient).Server = this;
+            client.Server = this;
         }
 
         /// <summary>
@@ -126,17 +126,17 @@ namespace Hellion.Cluster
             Log.Info("Loading configuration...");
 
             if (File.Exists(ClusterConfigurationFile) == false)
-                ConfigurationManager.Save(new LoginConfiguration(), ClusterConfigurationFile);
+                JsonHelper.Save(new ClusterConfiguration(), ClusterConfigurationFile);
 
-            this.ClusterConfiguration = ConfigurationManager.Load<ClusterConfiguration>(ClusterConfigurationFile);
+            this.ClusterConfiguration = JsonHelper.Load<ClusterConfiguration>(ClusterConfigurationFile);
 
             this.ServerConfiguration.Ip = this.ClusterConfiguration.Ip;
             this.ServerConfiguration.Port = this.ClusterConfiguration.Port;
 
             if (File.Exists(DatabaseConfigurationFile) == false)
-                ConfigurationManager.Save(new DatabaseConfiguration(), DatabaseConfigurationFile);
+                JsonHelper.Save(new DatabaseConfiguration(), DatabaseConfigurationFile);
 
-            this.DatabaseConfiguration = ConfigurationManager.Load<DatabaseConfiguration>(DatabaseConfigurationFile);
+            this.DatabaseConfiguration = JsonHelper.Load<DatabaseConfiguration>(DatabaseConfigurationFile);
 
             Log.Done("Configuration loaded!");
         }
@@ -176,7 +176,8 @@ namespace Hellion.Cluster
 
             try
             {
-                this.connector.Connect(this.ClusterConfiguration.ISC.Ip, this.ClusterConfiguration.ISC.Port);
+                var resolvedIp = HostResolver.ResolveToIp(this.ClusterConfiguration.ISC.Ip);
+                this.connector.Connect(resolvedIp, this.ClusterConfiguration.ISC.Port);
                 this.iscThread = new Thread(this.connector.Run);
                 this.iscThread.Start();
             }