+++ src/Hellion.Core/Data/Headers/InventoryParts.cs
@@ -1,0 +1,38 @@
+﻿namespace Hellion.Core.Data.Headers
+{
+    public class InventoryParts
+    {
+        public const int PARTS_HEAD = 0;
+        public const int PARTS_HAIR = 1;
+        public const int PARTS_UPPER_BODY = 2;
+        public const int PARTS_LOWER_BODY = 3;
+        public const int PARTS_HAND = 4;
+        public const int PARTS_FOOT = 5;
+        public const int PARTS_CAP = 6;
+        public const int PARTS_ROBE = 7;
+        public const int PARTS_CLOAK = 8;
+        public const int PARTS_LWEAPON = 9;
+        public const int PARTS_RWEAPON = 10;
+        public const int PARTS_SHIELD = 11;
+        public const int PARTS_MASK = 12;
+        public const int PARTS_RIDE = 13;
+        public const int PARTS_CAP2 = 14;
+        public const int PARTS_UPPER2 = 15;
+        public const int PARTS_LOWER2 = 16;
+        public const int PARTS_HAND2 = 17;
+        public const int PARTS_FOOT2 = 18;
+        public const int PARTS_NECKLACE1 = 19;
+        public const int PARTS_RING1 = 20;
+        public const int PARTS_RING2 = 21;
+        public const int PARTS_EARRING1 = 22;
+        public const int PARTS_EARRING2 = 23;
+        public const int PARTS_PROPERTY = 24;
+        public const int PARTS_BULLET = 25;
+        public const int PARTS_HAT = 26;
+        public const int PARTS_CLOTH = 27;
+        public const int PARTS_GLOVE = 28;
+        public const int PARTS_BOOTS = 29;
+        public const int PARTS_CLOAK2 = 30;
+        public const int MAX_HUMAN_PARTS = 31;
+    }
+}