+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 ## Hellion ChangeLog
 
+### Update #14
+
+Date: 18/03/2017
+
+- World
+  - Load JobData (propJob.inc)
+  - Damage formula
+  - Miss / Blocking / Critical hit calculation
+  - Knockback
+  - Add Item refine / element / element refine
+
+### Update #13
+
+Date: 11/03/2017
+
+- Common
+  - Fix configuration hostname issue [PR #42](https://github.com/Eastrall/Hellion/pull/42) (by Almewty)
+  - Remove unnecessary checks [Issue #44](https://github.com/Eastrall/Hellion/issues/44)
+  - Rename ConfigurationManager to JsonHelper [Issue #43](https://github.com/Eastrall/Hellion/issues/43)
+- World
+  - Fix/optimize new walk algorithm [Issue #36](https://github.com/Eastrall/Hellion/issues/36)
+
 ### Update #12
 
 Date: 04/03/2017